+++ CHANGELOG.md
@@ -1,17 +1,17 @@
 # Changelog
 All notable changes to this project will be documented in this file
 
-## [Unreleased]
-
 ## [0.2.0]
 ### Added
 - Version numbers above Copyright text
-- 
+- New ID for getting units as a type to convert from
+- New function to take units as convert from type to show units
 
 ### Changed
 - Body text to white
 - Header text to white
 - Answers/Results text to white
+- Copyright to Ghazi Design
 
 ## 0.0.1 [2017.06.30]
 ### Added